requirements.txt
discord.py
requests
pprint
dateutils